The RealTimeClock library leverages the on-board RTC available on the XMEGA, and provides services like allowing you to 	* set and query the time 	* set a periodic callback, to be triggered every N seconds
Along with modifications to the Xmegaduino library's wiring implementation, this will allow you to easily deal with an
RTC-based time system.  Use this library to get and set the real datetime, and use millis()/micros() as usual.

Assumptions:
	1) You are on a platform that has either an RTC or RTC32 support

	2) You are using a variant that actually enables the RTC
		(defines USE_RTC in the pins_arduino.h)

	3) You are using an Xmegaduino (or other Arduino compatible
		sdk) that has implemented my mods for the RTC wiring
		functionality (e.g. from https://github.com/psychogenic/Xmegaduino
		or from Xmegaduino main trunk, if the pull request has been merged).

Documentation is available within the source, on http://flyingcarsandstuff.com/projects/RealTimeClock and in the sample code included in the examples directory.

Pat Deegan, http://flyingcarsandstuff.com